+++ client/src/components/GarageSlider/GarageSlider.js
@@ -1,35 +1,65 @@
-import React, { useState, useEffect, useContext } from 'react';
+import React, { Component } from 'react';
 import { Carousel } from 'react-bootstrap';
+import SliderCaption from './SliderCaption';
+import ImageSlider from '../CarSlider/ImageSlider';
 
-import SliderCaption from './../SliderCaption/SliderCaption';
-
-import UserContext from './../../context/user-context';
-import GarageSliderContext from './../../context/garage-slider-context';
 import './GarageSlider.scss';
 
-const GarageSlider = () => {
-  const { cars } = useContext(UserContext);
-  const [index, setIndex] = useState(0);
-  const [direction, setDirection] = useState(null);
+class GarageSliderLogic extends Component {
+  constructor(props, context) {
+    super(props, context);
+    this.state = {
+      index: 0,
+      direction: null
+    };
+  }
 
-  const handleSelect = (selectedIndex, e) => {
-    setIndex(selectedIndex);
-    setDirection(e.direction);
+  handleSelect = (selectedIndex, e) => {
+    this.setState({
+      index: selectedIndex,
+      direction: e.direction
+    });
   };
 
-  const renderCars = cars.map(car => {
+  render() {
+    return this.props.children({
+      ...this.state,
+      handleSelect: this.handleSelect,
+      selectCurrentCar: this.selectCurrentCar
+    });
+  }
+}
+
+export default ({ user, selectCurrentCar }) => {
+  const renderCars = user.cars.map(car => {
     return (
-      <Carousel.Item key={car._id}>
-        <SliderCaption car={car} />
+      <Carousel.Item>
+        <SliderCaption
+          selectCurrentCar={selectCurrentCar}
+          imgSrc={car.mediumImg}
+          car={car}
+        />
+        <ImageSlider className="CarSlide" src={car.largeImg} />
       </Carousel.Item>
     );
   });
-
   return (
-    <div className="GarageSlider-container">
-      <Carousel>{renderCars}</Carousel>
-    </div>
+    <GarageSliderLogic>
+      {props => (
+        <div className="GarageSlider-container">
+          <Carousel
+            activeIndex={props.index}
+            direction={props.direction}
+            onSelect={props.handleSelect}
+            indicators={false}
+            fade={false}
+            controls={true}
+            interval="5000"
+          >
+            {renderCars}
+          </Carousel>
+        </div>
+      )}
+    </GarageSliderLogic>
   );
 };
-
-export { GarageSlider as default };
