+++ client/src/containers/Garage/Garage.js
@@ -1,45 +1,23 @@
-import React, { useState, useEffect, useReducer } from 'react';
+import React from 'react';
 
-import Header from './../../components/Header/Header';
-import GarageSlider from './../../components/GarageSlider/GarageSlider'
-// Helper functions
-import validateToken from './../../lib/userValidation';
-import userReducer from './../../reducers/users';
-import UserContext from './../../context/user-context';
-import instance from './../../lib/axios';
+import GarageSlider from '../../components/GarageSlider/GarageSlider';
+import Header from '../../components/Header/Header';
+import GarageLogic from './GarageLogic';
+import GarageModal from '../../components/GarageSlider/GarageModal';
 
-const Garage = () => {
-  const [user, userDispatch] = useReducer(userReducer, {
-    username: 'Guest',
-    email: null
-  });
-  const [cars, setCars] = useState([]);
-
-  useEffect(() => {
-    validateToken(userDispatch);
-  }, []);
-
-  useEffect(() => {
-    const getUserCars = async () => {
-      try {
-        const { data } = await instance.get('/cars/user');
-        const userCars = data.cars
-
-        setCars(userCars);
-      } catch (e) {
-        console.log(e.message);
-      }
-    };
-
-    getUserCars();
-  }, [user]);
-
-  return (
-    <UserContext.Provider value={{ user, cars }}>
-      <Header />
-      <GarageSlider />
-    </UserContext.Provider>
-  );
-};
-
-export { Garage as default };
+export default () => (
+  <GarageLogic>
+    {props => (
+      <div>
+        <Header user={props.user} />
+        <div>
+          <GarageSlider
+            user={props.user}
+            selectCurrentCar={props.selectCurrentCar}
+          />
+          <GarageModal show={props.show} />
+        </div>
+      </div>
+    )}
+  </GarageLogic>
+);
