+++ client/src/containers/Garage/GarageLogic.js
@@ -1,7 +1,7 @@
 import React, { Component} from 'react';
 
-import tokenValidationHelper from '../../../lib/tokenValidationHelper';
-import instance from '../../../lib/axios';
+import tokenValidationHelper from '../../lib/tokenValidationHelper';
+import instance from '../../lib/axios';
 
 export default class Garage extends Component {
   constructor() {
